orchestrator/subagents/cropDoctor/requirements.txt
requests
geopy
pillow
google-generativeai
numpy
google-adk
ipython
vertexai
fastapi
uvicorn
pydantic